Reduce memory consumption.
Storing short strings in a python list results in a huge overhead for each
stored item.  Instead, pack lots of them into a bytearray so that the
overhead becomes insignificant.

To reduce the space needed while sorting, bin the data into 256 bins
using the top 8 bits of mer_id.  As long as they are divided fairly
evenly, this divides the memory used by the sort function by 256.  The
sort function itself is implemented as an LSD radix sort on the packed
bytearrays.  This actually runs faster than the standard python sort
presumably because fewer data manipulations are needed.  It also only
needs memory of the order of the input data as opposed to the sorted
function which returns a list will all of the overheads that implies.
Unfortunately the sort can't be easily done in-place as it needs to be
stable.

Change the option that limits the amount sorted in each chunk to -m
(--memlimit) in megabytes.  This makes it easier to guess how much
memory will actually be used.